backend/requirements.txt
flask
flask-cors
python-dotenv
ollama
chromadb
sentence-transformers